+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Each tagged release MUST have a section 2 heading starting at the time of releas
   - Added tab completion when `MEDIAWIKI_VOLUMES_CODE` is entered via wizard
   - Added `docker update` command to pull and update all created containers
   - Added `docker hosts where` command to show you where the hosts file is
+  - Added `docker <service> expose` command for most services, exposing an internal port locally (T299514)
   - Added the ability to run multiple seperate development environments via the `--context` flag (T301002)
   - Added `docker mediawiki get-code` command for fetching MediaWiki, skins and extensions from Gerrit
   - Removed code clone wizard from `docker mediawiki` command startup, insteading prompting users to use `docker mediawiki get-code`